Separate tests for counters
